+++ src/main/java/com/daou/sabangnetserver/repository/OrdersDetailRepository.java
@@ -3,5 +3,5 @@ package com.daou.sabangnetserver.repository;
 import com.daou.sabangnetserver.model.OrdersDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrdersDetailRepository extends JpaRepository<OrdersDetail, Integer> {
+public interface OrdersDetailRepository extends JpaRepository<OrdersDetail, Long> {
 }
